+++ test/triggers/document-generated.test.js
@@ -1,14 +1,13 @@
 'use strict';
 
 const App = require('../../index');
-const documentSample = require('../../samples/document');
+const documentCardSample = require('../../samples/document-card');
 const zapier = require('zapier-platform-core');
 const { bundleWithAuth, pdfmonkeyApi } = require('../test-helpers');
 
 const appTester = zapier.createAppTester(App);
 const processesDocumentSample = {
-  ...documentSample,
-  parsedPayload: { name: 'Jane Doe' },
+  ...documentCardSample,
   parsedMeta: { _filename: 'demo-document.pdf' }
 };
 
@@ -18,12 +17,12 @@ describe('Triggers::DocumentGenerated', () => {
   describe('when only the Workspace is specified', () => {
     it('returns a list of Documents', (done) => {
       pdfmonkeyApi
-        .get('/api/v1/documents')
+        .get('/api/v1/document_cards')
         .query({
           'q[workspace_id]': '11111111-2222-3333-4444-555555555555',
           'page[size]': 100
         })
-        .reply(200, { documents: [documentSample] });
+        .reply(200, { document_cards: [documentCardSample] });
 
       const bundle = {
         ...bundleWithAuth(),
@@ -44,13 +43,13 @@ describe('Triggers::DocumentGenerated', () => {
   describe('when both the Workspace and Template are specified', () => {
     it('returns a list of Documents', (done) => {
       pdfmonkeyApi
-        .get('/api/v1/documents')
+        .get('/api/v1/document_cards')
         .query({
           'q[workspace_id]': '11111111-2222-3333-4444-555555555555',
           'q[document_template_id]': '22222222-3333-4444-5555-666666666666,33333333-4444-5555-6666-777777777777',
           'page[size]': 100
         })
-        .reply(200, { documents: [documentSample] });
+        .reply(200, { document_cards: [documentCardSample] });
 
       const bundle = {
         ...bundleWithAuth(),